Fix issue introduced by 05e3289
I had a lot of lines in the NAT like
[translate_addr: 144] mtcp_getpeer() failed sock=61 side=1

In nat.c, MOS_SIDE_CLI has been changed by MOS_SIDE_BOTH, without
changing the length of the sockaddr len.
Actually, by luck up to this patch the size of a pointer * 2 was equal
to the size of a sockaddr, but now the error in the sizeof appears.

I suspect this also happens in midstat.c, but I don't have time to push
this further.